Consolidate story list layouts.
