+++ clients/android/NewsBlur/res/layout/row_folderitem.xml
@@ -15,80 +15,67 @@
         android:layout_height="match_parent"
         android:layout_toRightOf="@id/row_item_favicon_borderbar_1" />
 
+    <View
+        android:id="@+id/row_item_sidebar"
+        android:layout_width="9dp"
+        android:layout_height="9dp"
+        android:layout_toRightOf="@id/row_item_favicon_borderbar_2"
+        android:layout_marginLeft="7dp"
+        android:layout_marginRight="7dp"
+        android:layout_centerVertical="true" />
+
     <ImageView
         android:id="@+id/row_item_feedicon"
         android:layout_width="18dp"
         android:layout_height="18dp"
-        android:layout_marginTop="6dp"
-        android:layout_marginLeft="22dp"
-        android:layout_toRightOf="@id/row_item_favicon_borderbar_2" />
+        android:layout_marginTop="4dp"
+        android:layout_toRightOf="@id/row_item_sidebar" />
 
     <TextView
         android:id="@+id/row_item_feedtitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginLeft="6dp"
-        android:layout_marginTop="5dp"
+        android:layout_marginTop="4dp"
         android:layout_toRightOf="@id/row_item_feedicon"
         android:ellipsize="end"
         android:singleLine="true"
         android:textColor="@color/lightgray"
         android:textSize="14dp" />
 
-    <RelativeLayout
-        android:id="@+id/row_item_title_container"
-        android:layout_width="match_parent"
+    <TextView
+        android:id="@+id/row_item_title"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_toRightOf="@id/row_item_favicon_borderbar_2"
-        android:layout_marginTop="4dp"
-        android:layout_below="@id/row_item_feedicon">
+        android:layout_toRightOf="@id/row_item_sidebar"
+        android:layout_below="@id/row_item_feedicon"
+        android:paddingTop="6dp"
+        android:paddingBottom="4dp"
+        android:paddingRight="4dp"
+        android:maxLines="2"
+        android:ellipsize="end"
+        android:textSize="14sp" />
 
-
-        <TextView
-            android:id="@+id/row_item_sidebar"
-            android:layout_width="9dp"
-            android:layout_height="9dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginLeft="7dp"
-            android:layout_marginRight="7dp"
-            android:layout_centerVertical="true"
-            android:layout_alignParentLeft="true"/>
-
-        <TextView
-            android:id="@+id/row_item_title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_toRightOf="@+id/row_item_sidebar"
-            android:layout_marginRight="8dp"
-            android:paddingTop="4dp"
-            android:paddingBottom="4dp"
-            android:maxLines="2"
-            android:ellipsize="end"
-            android:textSize="14sp"/>
-
-        <TextView
-            android:id="@+id/row_item_content"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_toRightOf="@+id/row_item_sidebar"
-            android:layout_below="@+id/row_item_title"
-            android:layout_marginRight="8dp"
-            android:paddingTop="0dp"
-            android:paddingBottom="4dp"
-            android:maxLines="2"
-            android:ellipsize="end"
-            android:textSize="14sp"/>
-
-    </RelativeLayout>
+    <TextView
+        android:id="@+id/row_item_content"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_toRightOf="@+id/row_item_sidebar"
+        android:layout_below="@+id/row_item_title"
+        android:paddingBottom="4dp"
+        android:paddingRight="4dp"
+        android:maxLines="2"
+        android:ellipsize="end"
+        android:textSize="14sp"/>
 
     <TextView
         android:id="@+id/row_item_date"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignParentRight="true"
-        android:layout_below="@+id/row_item_title_container"
-        android:layout_marginRight="12dp"
+        android:layout_below="@id/row_item_content"
         android:paddingBottom="4dp"
+        android:paddingRight="8dp"
         android:textColor="@color/darkgray"
         android:textSize="11dp" />
 
@@ -96,13 +83,14 @@
         android:id="@+id/row_item_author"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_below="@+id/row_item_title_container"
-        android:paddingBottom="4dp"
+        android:layout_below="@id/row_item_content"
+        android:layout_toRightOf="@id/row_item_sidebar"
         android:layout_toLeftOf="@id/row_item_date"
-        android:layout_toRightOf="@id/row_item_favicon_borderbar_2"
-        android:layout_marginLeft="22dp"
+        android:layout_alignLeft="@id/row_item_content"
+        android:paddingBottom="4dp"
         android:ellipsize="end"
         android:singleLine="true"
         android:textColor="@color/darkgray"
-        android:textSize="11dp" />
+        android:textSize="11sp" />
+
 </RelativeLayout>
