--- clients/android/NewsBlur/res/layout/row_socialitem.xml
@@ -1,103 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    style="?selectorStoryBackground"
-    android:orientation="horizontal" >
-
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent" >
-
-        <View
-            android:id="@+id/row_item_favicon_borderbar_1"
-            android:layout_width="8dp"
-            android:layout_height="match_parent" />
-
-        <View
-            android:id="@+id/row_item_favicon_borderbar_2"
-            android:layout_width="8dp"
-            android:layout_height="match_parent"
-            android:layout_toRightOf="@id/row_item_favicon_borderbar_1" />
-
-        <TextView
-            android:id="@+id/row_item_sidebar"
-            android:layout_width="10dp"
-            android:layout_height="10dp"
-            android:layout_centerVertical="true"
-            android:layout_marginLeft="5dp"
-            android:layout_marginRight="5dp"
-            android:layout_toRightOf="@id/row_item_favicon_borderbar_2" />
-
-        <ImageView
-            android:id="@+id/row_item_feedicon"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:layout_marginTop="10dp"
-            android:layout_toRightOf="@id/row_item_sidebar" />
-
-        <TextView
-            android:id="@+id/row_item_feedtitle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="8dp"
-            android:layout_toRightOf="@id/row_item_feedicon"
-            android:layout_alignBottom="@id/row_item_feedicon"
-            android:ellipsize="end"
-            android:singleLine="true"
-            android:textColor="@color/darkgray"
-            android:textSize="14sp" />
-
-        <TextView
-                android:id="@+id/row_item_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignLeft="@id/row_item_feedicon"
-                android:layout_below="@id/row_item_feedicon"
-                android:layout_marginRight="8dp"
-                android:paddingTop="4dp"
-                android:paddingBottom="2dp"
-                android:textColor="@color/darkgray"
-                android:textSize="14sp"/>
-
-        <TextView
-                android:id="@+id/row_item_content"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignLeft="@id/row_item_feedicon"
-                android:layout_below="@id/row_item_title"
-                android:layout_marginRight="8dp"
-                android:paddingTop="2dp"
-                android:paddingBottom="4dp"
-                android:maxLines="2"
-                android:ellipsize="end"
-                android:textColor="@color/darkgray"
-                android:textSize="14sp"/>
-        
-        <TextView
-            android:id="@+id/row_item_date"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentRight="true"
-            android:layout_below="@+id/row_item_content"
-            android:layout_marginRight="8dp"
-            android:paddingBottom="4dp"
-            android:textColor="@color/darkgray"
-            android:textSize="11dp" />
-
-        <TextView
-            android:id="@+id/row_item_author"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/row_item_content"
-            android:paddingBottom="4dp"
-            android:layout_toLeftOf="@id/row_item_date"
-            android:layout_toRightOf="@id/row_item_sidebar"
-            android:ellipsize="end"
-            android:singleLine="true"
-            android:textColor="@color/darkgray"
-            android:textSize="11dp" />
-        
-    </RelativeLayout>
-
-</LinearLayout>
