+++ clients/android/NewsBlur/src/com/newsblur/database/StoryItemsAdapter.java
@@ -8,13 +8,31 @@ import java.util.List;
 
 import com.newsblur.domain.Story;
 
-public abstract class StoryItemsAdapter extends SimpleCursorAdapter {
+public class StoryItemsAdapter extends SimpleCursorAdapter {
 
-	public StoryItemsAdapter(Context context, int layout, Cursor c, String[] from, int[] to) {
+	protected Cursor cursor;
+
+    // should be subclassed to handle bindView(), not created directly
+	protected StoryItemsAdapter(Context context, int layout, Cursor c, String[] from, int[] to) {
         // don't set *any* flags, we use auto-refreshing Loaders or plain Loaders and an explict load after syncs
 		super(context, layout, c, from, to, 0);
+        cursor = c;
     }
 
-    public abstract Story getStory(int position);
+	@Override
+	public int getCount() {
+		return cursor.getCount();
+	}
+
+	@Override
+	public Cursor swapCursor(Cursor c) {
+		this.cursor = c;
+		return super.swapCursor(c);
+	}
+
+	public Story getStory(int position) {
+        cursor.moveToPosition(position);
+        return Story.fromCursor(cursor);
+    }
 
 }
