+++ clients/android/NewsBlur/src/com/newsblur/fragment/ReadStoriesItemListFragment.java
@@ -40,7 +40,7 @@ public class ReadStoriesItemListFragment extends ItemListFragment {
         if ((adapter == null) && (cursor != null)) {
             String[] groupFrom = new String[] { DatabaseConstants.STORY_TITLE, DatabaseConstants.STORY_SHORT_CONTENT, DatabaseConstants.STORY_AUTHORS, DatabaseConstants.STORY_TIMESTAMP, DatabaseConstants.SUM_STORY_TOTAL, DatabaseConstants.FEED_TITLE };
             int[] groupTo = new int[] { R.id.row_item_title, R.id.row_item_content, R.id.row_item_author, R.id.row_item_date, R.id.row_item_sidebar, R.id.row_item_feedtitle };
-            adapter = new MultipleFeedItemsAdapter(getActivity(), R.layout.row_socialitem, cursor, groupFrom, groupTo);
+            adapter = new MultipleFeedItemsAdapter(getActivity(), R.layout.row_folderitem, cursor, groupFrom, groupTo);
             adapter.setViewBinder(new SocialItemViewBinder(getActivity(), true));
             itemList.setAdapter(adapter);
        }
